+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 2.20.1 / aria branch / Unreleased
+  * [Fix] #659  handling of function keys, when quicksearch is on
   * [Improved] #656 WAI-ARIA support
     - Set focus to first node on first tab-in
     - Support [home] and [end] keys